refactor: changing the template structure
